sprint-boot-mvc/src/main/java/org/november/swimmer/GreetingController.java
package org.november.swimmer;

import com.sun.deploy.net.HttpResponse;
import org.springframework.http.HttpRequest;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;

import javax.servlet.http.HttpSession;

@Controller
public class GreetingController {

    @RequestMapping("/greeting")
    public String greeting(@RequestParam(value="name", required = false, defaultValue = "World") String name, Model model) {
        model.addAttribute("name", name);
        return "greeting";
    }

    @RequestMapping("/greetingSession")
    public String greetingSession(Model model, HttpSession session,
                                  @RequestParam(value="name", required = false, defaultValue = "World") String name){
        session.setAttribute("name", name);

        return "greeting";
    }
}
